[b2r2lib] Add handling of other events than POLLIN
In the callback thread a poll is performed on the blt handle
in order to know if the thread should be terminated or not.

When poll returns OK (>0) the event field can contain other events
than POLLIN. If another event has occurred the thread has to be
terminated because it is likely that the blt handle is closed.

ST-Ericsson ID: 350511, 342717, 350544
ST-Ericsson FOSS-OUT ID: Trivial
Signed-off-by: Jimmy Rubin <jimmy.rubin@stericsson.com>

Change-Id: I289d1dcd68398c3c1ba489785b24289455304e58
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/26664
Reviewed-by: QATOOLS
Reviewed-by: QATEST
Tested-by: Jimmy RUBIN <jimmy.rubin@stericsson.com>
Reviewed-by: Per-Daniel OLSSON <per-daniel.olsson@stericsson.com>
Reviewed-by: Robert LIND <robert.lind@stericsson.com>