Add background blend support
Change API struct to correspond to kernel b2r2 driver updates.
Instead of blending only on the destination there is now
a possibility to blend on a supplied background using standard
blending flags. The background blending option is enabled through
setting a flag in the request. The destination blending will in
this case be entirely omitted and the destination buffer is
overwritten with the result of the source and background blend
operation.

ST-Ericsson Linux next: NA
Depends-On: Ifeb8eb66c2fed97c541b09715a17251a81e493ab
ST-Ericsson ID: 350337
ST-Ericsson FOSS-OUT ID: Trivial

Introduce non-backward compatible changes to exposed interfaces
Signed-off-by: Jörgen Nilsson <jorgen.nilsson@stericsson.com>

Change-Id: Ia34f98cfb20e2f0486049e82581dcea28b465ed4
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/34025
Tested-by: Jorgen NILSSON <jorgen.nilsson@stericsson.com>
Reviewed-by: Jimmy RUBIN <jimmy.rubin@stericsson.com>
Reviewed-by: QATOOLS
Reviewed-by: Robert LIND <robert.lind@stericsson.com>